+++ packages/dds/sequence/src/intervalCollectionMapInterfaces.ts
